Bonooru's Song Sync
An attempt at syncing the Scarecrow's Song for players who have yet to visit and do so themselves. Incomplete due to missing emits for notifications and a proper check for the relevant flag for the Adult half of the process. Relies on changes to ML64 in its own commits on its own branch.