Some fixes
Every cur.permSceneDataStructureField[i] was coming up as undefined, making the core update everything all the time as it never matched. Since I used it as the template for the Scarecrow stuff, it too was wrong. Implemented one-time-syncing, though it's still not functioning properly for Client players. Need to find some other place the Scarecrow Song is being stored to outside of the known 0x1450 save data range?