Mish-mash of Stuff from Work-in-Progress (#25)
* Maybe fix Water Temple

I did this in the wrong branch damnit.

* Cherry-picked & Modified Water Temple Switch Fix

Moving the conditions to the area where the permSceneData's Water Temple switches are being checked in the first place.

* Fix + Tracking Deku Theater Flags

Some checks, though the Mask of Truth's is redundant unless the game's been modified to have a speedy Bunny Hood.

* Disabling Buggy MW Notif

Current behavior of Notifs appears to only execute it for the Server+Client, annoyingly, and only produces notifications when the host collects items for other players' worlds' on the host's instance.

* Switching execution order around

This should still work even if the save detected isn't Vanilla or Randomized.

* Fix notification sound bug fix

Notification sounds shouldn't be playing if config.notificationSound's been set to false at all, right?

* Re-add code forcing max 20 heart containers

Just in case something somehow goes wrong, or an outdated or modified save is used to host a lobby.

* Add Fast Bunny Hood RomFlag

What it says on the tin: Figured out how to check for Bunny Hood tampering on the Randomizer's side. Also expanded the version check, just in case people come in using unstable builds of the rando in the future.

* Attempt to replace deprecated flag_fixer.ovl

Not being able to check exactly what flag_fixer.ovl does, I tried to make a somewhat uneducated guess here. If this looks good, it could lead to finally removing the flag_fixer.ovl code in OotOnlineClient.ts, I suppose?