Fix OotR ice trap loop
This fixes an issue introduced by TestRunnerSRL/OoT-Randomizer#1555 where a 'pending' value for freezing Link is stored in an unused uint32 of the Fishing Pond.

OotO syncs all values that are part of the scene flags, including the unused ones, due to rando storing data there like flags for scrubs.

This commit simply adds a bit mask for blacklisting stuff from the scene flags and blacklists the Fishing Pond's unused data.